Add Android SDK cmake 3.10.2 component to ubuntu16.04, 18.04 and windows.
Android SDK gradle plugin supports fixed versions of cmake, which are
either 3.6.4111459 or 3.10.2. 4988404, and on Github Actions our apps
fail to build if their build.gradle specify cmake 3.10.2 because
Android Gradle Plugin tries to download and install it if it's missing
(and causes file write access violation, at least on Ubuntu).